requirements.txt
pymongo>=4.6.0
pandas>=2.0.0
dnspython>=2.4.0
python-dotenv>=1.0.0
certifi>=2023.7.22
urllib3>=2.0.0